refactor: move selection popup to layout
